+++ back-end/src/modules/orphanages/services/DeleteOrphanageService.ts
@@ -13,20 +13,22 @@ class DeleteOrphanageService {
     private orphanageRepository: IOrphanagesRepository,
   ) {}
 
-  execute = async (id: string) => {
+  execute = async (id: string, user_id: number) => {
     const checkOrphanage = await this.orphanageRepository.findById(id);
 
     if (!checkOrphanage) throw new AppError('Orphanage not found.');
 
-    await this.orphanageRepository.delete(checkOrphanage);
+    if (checkOrphanage.user.id !== user_id)
+      throw new AppError('You cannot delete this orphanage');
 
     if (checkOrphanage.images) {
-      checkOrphanage.images.forEach(async image => {
+      checkOrphanage.images.forEach(image => {
         const pathFile = path.resolve(folder, image.path);
-
+        if (!fs.existsSync(pathFile)) throw new AppError('Image not found');
         fs.unlinkSync(pathFile);
       });
     }
+    await this.orphanageRepository.delete(checkOrphanage);
   };
 }
